More Symlink Nonsense

+++ Makefile
@@ -1,5 +1,5 @@
 build:
-	ln -s /localstorage
+	[ ! -f ./localstorage ] && ln -s /localstorage
 	for container in beam scaler; do \
 		mkdir -p $$container/assets ; \
 		sed "s=XAUTHKEY=$$(xauth list | head -n 1)=" $$container/templates/entrypoint.sh > $$container/assets/entrypoint.sh ; \
